rename state machine file
